Remove `FormMetadataInterface`
